SAR_waterbody_detection - 副本/requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
efficientnet_pytorch==0.7.1
einops==0.8.0
filelock==3.16.1
fsspec==2024.10.0
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
monai==1.4.0
mpmath==1.3.0
munch==4.0.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pretrainedmodels==0.7.4
python-dateutil==2.9.0.post0
pytorch-triton==3.1.0+cf34004b8a
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
safetensors==0.4.5
segmentation-models-pytorch==0.3.4
setuptools==70.2.0
six==1.16.0
sympy==1.13.1
timm==0.9.7
torch==2.6.0.dev20241115+cu124
torchaudio==2.5.0.dev20241116+cu124
torchvision==0.20.0.dev20241116+cu124
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
